refactor: nav code cleanup
