+++ src/theme/Navbar/MobileNavItem.jsx
@@ -1,0 +1,46 @@
+import React from 'react';
+import clsx from 'clsx';
+
+import NavLink from './NavLink';
+
+const  MobileNavItem = ({ items, className, ...props }) => {
+  // Need to destructure position from props so that it doesn't get passed on.
+  const navLinkClassNames = (extraClassName, isSubList = false) =>
+    clsx(
+      'menu__link',
+      {
+        'menu__link--sublist': isSubList,
+      },
+      extraClassName
+    );
+
+  if (!items) {
+    return (
+      <li className="menu__list-item">
+        <NavLink className={navLinkClassNames(className)} {...props} />
+      </li>
+    );
+  }
+
+  return (
+    <li className="menu__list-item">
+      <NavLink className={navLinkClassNames(className, true)} {...props}>
+        {props.label}
+      </NavLink>
+      <ul className="menu__list">
+        {items.map(({ className: childItemClassName, ...childItemProps }, i) => (
+          <li className="menu__list-item" key={i}>
+            <NavLink
+              activeClassName="menu__link--active"
+              className={navLinkClassNames(childItemClassName)}
+              {...childItemProps}
+              onClick={props.onClick}
+            />
+          </li>
+        ))}
+      </ul>
+    </li>
+  );
+};
+
+export default MobileNavItem;
