+++ src/theme/Navbar/styles.module.css
@@ -5,12 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-@media screen and (max-width: 997px) {
-  .displayOnlyInLargeViewport {
-    display: none !important;
-  }
-}
-
 .navbarHideable {
   transition: top 0.2s ease-in-out;
 }
@@ -39,16 +33,16 @@
 }
 
 .navbarLogoCustom {
-  margin-top: 0.5rem;
   height: 1.5rem;
-  margin-left: 0.33rem;
-}
-
-.navbarLogoTextCustom {
-  height: 1.5rem;
+	vertical-align: middle;
   margin-left: 0.33rem;
 }
 
 .navbarWhite {
   background-color: inherit;
 }
+
+.navbarMenu {
+	cursor: pointer;
+}
+
